requirements.txt
numpy==1.9.1
pyserial==2.7
scipy==0.14.0
matplotlib==1.4.2
